2024103391韩璐瑶/结课项目/protein_charge_analysis/requirements.txt
numpy==1.26.4
matplotlib==3.8.2
Jinja2==3.1.3
pandas==2.2.0
biopython==1.83
scipy==1.12.0